+++ R/zzz.R
@@ -8,7 +8,8 @@ if (getRversion() >= "2.15.1") {
     "partial_dependency", "rowid", "se", "resampled_set", "observations", 
     "model", "predictions", "COE", "MB", "NMB", "RMSE", "default",
     "normalised_root_mean_squared_error", "r", "r_squared", 
-    "root_mean_squared_error", "partial_dependencies", "IOA", "MGE", "NMGE"
+    "root_mean_squared_error", "partial_dependencies", "IOA", "MGE", "NMGE",
+    "year_met", "date_unix", "n_sample"
   )
   
   # Squash the notes