+++ R/base_re_exports.R
@@ -1,4 +1,6 @@
 #' Pseudo-function to re-export functions from the \strong{stats} package. 
+#' 
+#' @name base functions
 #'
 #' @importFrom stats median na.omit predict reorder setNames
 NULL