+++ R/dplyr_re_exports.R
@@ -1,4 +1,6 @@
 #' Pseudo-function to re-export \strong{dplyr}'s common functions. 
+#' 
+#' @name dplyr functions
 #'
 #' @importFrom dplyr select rename mutate filter arrange distinct summarise 
 #'     do group_by ungroup rowwise do left_join inner_join everything bind_rows