+++ R/zzz.R
@@ -1,5 +1,7 @@
 #' Squash the global variable notes when building a package. 
 #' 
+#' @name zzz
+#' 
 if (getRversion() >= "2.15.1") {
   
   # What variables are causing issues?